- Added new isConnected() function. Will return TRUE if the row is connected to a live table. - Refactored setTable() and getTable() functions. Implemented a default and forced contructor to call setTable() - Fixed issue with getTable() called before the constructor of KDatabaseRowTable ran call through the reset() fucntion. Solved by checking if the table is connected and added extra reset() code to the constructor of KDatabaseTableRow - Added isConnected checks to all functions to prevent operations on the table is the row object is not connected.

+++ code/libraries/koowa/database/row/table.php
@@ -24,7 +24,7 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 *
 	 * @var	string|object
 	 */
-	protected $_table;
+	protected $_table = false;
 
 	/**
      * Object constructor 
@@ -35,10 +35,15 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	{
 		parent::__construct($config);
 
-		// Set the table indentifier
-		if(isset($config->table)) {
-			$this->setTable($config->table);
-		}
+		$this->setTable($config->table);
+			
+		// Reset the row
+        $this->reset();
+            
+        // Reset the row data
+        if(isset($config->data))  {
+            $this->setData($config->data->toArray(), $this->_new);
+        }
 	}
 
 	/**
@@ -52,31 +57,37 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	protected function _initialize(KConfig $config)
 	{
 		$config->append(array(
-			'table'	=> null
+			'table'	=> $this->_identifier->name
 		));
 
 		parent::_initialize($config);
 	}
 
 	/**
-	 * Get the identifier for the table with the same name
-	 *
-	 * @return	KIdentifierInterface
-	 */
-	public function getTable()
-	{
-		if(!$this->_table)
-		{
-			$identifier 		= clone $this->_identifier;
-			$identifier->name	= KInflector::tableize($identifier->name);
-			$identifier->path	= array('database', 'table');
+     * Method to get a table object
+     * 
+     * Function catches KDatabaseTableExceptions that are thrown for tables that 
+     * don't exist. If no table object can be created the function will return FALSE.
+     *
+     * @return KDatabaseTableAbstract
+     */
+    public function getTable()
+    {
+        if($this->_table !== false)
+        {
+            if(!($this->_table instanceof KDatabaseTableAbstract))
+		    {   		        
+		        try {
+		            $this->_table = KFactory::get($this->_table);
+                } catch (KDatabaseTableException $e) {
+                    $this->_table = false;
+                }
+            }
+        }
 
-			$this->_table = KFactory::get($identifier);
-		}
-
-		return $this->_table;
-	}
-
+        return $this->_table;
+    }
+	
 	/**
 	 * Method to set a table object attached to the rowset
 	 *
@@ -85,21 +96,38 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 * @throws	KDatabaseRowException	If the identifier is not a table identifier
 	 * @return	KDatabaseRowsetAbstract
 	 */
-	public function setTable($table)
+    public function setTable($table)
 	{
 		if(!($table instanceof KDatabaseTableAbstract))
 		{
-			$identifier = KFactory::identify($table);
-
-			if($identifier->path[0] != 'table') {
-				throw new KModelException('Identifier: '.$identifier.' is not a table identifier');
+			if(is_string($table) && strpos($table, '.') === false ) 
+		    {
+		        $identifier         = clone $this->_identifier;
+		        $identifier->path   = array('database', 'table');
+		        $identifier->name   = KInflector::tableize($table);
+		    }
+		    else  $identifier = KFactory::identify($table);
+		    
+			if($identifier->path[1] != 'table') {
+				throw new KDatabaseRowsetException('Identifier: '.$identifier.' is not a table identifier');
 			}
 
-			$table = KFactory::get($identifier);
+			$table = $identifier;
 		}
 
 		$this->_table = $table;
+
 		return $this;
+	}
+	
+	/**
+	 * Test the connected status of the row.
+	 *
+	 * @return	boolean	Returns TRUE if we have a reference to a live KDatabaseTableAbstract object.
+	 */
+    public function isConnected()
+	{
+	    return (bool) $this->getTable();
 	}
 
 	/**
@@ -113,22 +141,23 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 		
 		if($this->_new)
 		{
-            $table = $this->getTable();
+            if($this->isConnected())
+            {
+                $data  = $this->getTable()->filter($this->getData(true), true);
+		        $row   = $this->getTable()->select($data, KDatabase::FETCH_ROW);
 
-		    $data  = $table->filter($this->getData(true), true);
-		    $row   = $table->select($data, KDatabase::FETCH_ROW);
-
-		    // Set the data if the row was loaded succesfully.
-		    if(!$row->isNew())
-		    {
-			    $this->setData($row->toArray(), false);
-			    $this->_modified = array();
-			    $this->_new      = false;
+		        // Set the data if the row was loaded succesfully.
+		        if(!$row->isNew())
+		        {
+			        $this->setData($row->toArray(), false);
+			        $this->_modified = array();
+			        $this->_new      = false;
 			    
-			    $this->setStatus(KDatabase::STATUS_LOADED);
-			    $result = true;
-		    }
-		    else $this->setStatus(KDatabase::STATUS_FAILED);
+			        $this->setStatus(KDatabase::STATUS_LOADED);
+			        $result = true;
+		        }
+		        else $this->setStatus(KDatabase::STATUS_FAILED);
+            }
 		}
 	
 		return $result;
@@ -144,28 +173,33 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function save()
 	{
-	    if($this->_new) 
+	    $result = false;
+	    
+	    if($this->isConnected())
 	    {
-		    $result = $this->getTable()->insert($this);
+	        if($this->_new) 
+	        {
+		        $result = $this->getTable()->insert($this);
 		    
-		    if($result !== false) {
-                $this->setStatus(KDatabase::STATUS_INSERTED);
-            } else {
-                $this->setStatus(KDatabase::STATUS_FAILED); 
-            }
-		} 
-		else 
-		{
-			$result = $this->getTable()->update($this);
+		        if($result !== false) {
+                    $this->setStatus(KDatabase::STATUS_INSERTED);
+                } else {
+                    $this->setStatus(KDatabase::STATUS_FAILED); 
+                }
+		    } 
+		    else 
+		    {
+			    $result = $this->getTable()->update($this);
 			
-			if($result > 0 ) {
-			    $this->setStatus(KDatabase::STATUS_UPDATED);
-			}
+			    if($result > 0 ) {
+			        $this->setStatus(KDatabase::STATUS_UPDATED);
+			    }
             
-			if($result <= 0) {
-			    $this->setStatus(KDatabase::STATUS_FAILED);
-			}
-		}
+			    if($result <= 0) {
+			        $this->setStatus(KDatabase::STATUS_FAILED);
+			    }
+		    }
+	    }
 
 		return (bool) $result;
     }
@@ -178,21 +212,24 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	public function delete()
 	{
 		$result = false;
-
-		if(!$this->_new) 
+		
+		if($this->isConnected())
 		{
-		    $result = $this->getTable()->delete($this);
+            if(!$this->_new) 
+		    {
+		        $result = $this->getTable()->delete($this);
 		    
-		    if($result !== false)
-            {
-                if($result > 0) {
-                    $this->setStatus(KDatabase::STATUS_DELETED);
-                }
+		        if($result !== false)
+                {
+                    if($result > 0) {
+                        $this->setStatus(KDatabase::STATUS_DELETED);
+                    }
             
-                if($result <= 0) {
-                    $this->setStatus(KDatabase::STATUS_FAILED);
+                    if($result <= 0) {
+                        $this->setStatus(KDatabase::STATUS_FAILED);
+                    }
                 }
-            }
+		    }
 		}
 
 		return (bool) $result;
@@ -206,10 +243,14 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	public function reset()
 	{
 		$result = false;
-	    
-	    if($this->_data = $this->getTable()->getDefaults()) {
-		    $this->setStatus(null);
-		    $result = true;
+		
+		if($this->isConnected())
+		{
+	        if($this->_data = $this->getTable()->getDefaults()) 
+	        {
+		        $this->setStatus(null);
+		        $result = true;
+		    }
 		}
 		
 		return $result;
@@ -222,10 +263,15 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function count()
 	{
-		$data  = $this->getTable()->filter($this->getData(true), true);
-		$count =  $this->getTable()->count($data);
+		$result = false;
+	    
+	    if($this->isConnected())
+		{
+	        $data   = $this->getTable()->filter($this->getData(true), true);
+		    $result = $this->getTable()->count($data);
+		}
 
-		return $count;
+		return $result;
 	}
 
 	/**
@@ -239,12 +285,15 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	 */
 	public function __unset($column)
 	{
-		$field = $this->getTable()->getColumn($column);
+		if($this->isConnected())
+		{
+	        $field = $this->getTable()->getColumn($column);
 
-		if(isset($field) && $field->required) {
-			$this->_data[$column] = $field->default;
-		} else {
-			parent::__unset($column);
+		    if(isset($field) && $field->required) {
+			    $this->_data[$column] = $field->default;
+		    } else {
+			    parent::__unset($column);
+		    }
 		}
 	}
 
@@ -263,7 +312,7 @@ class KDatabaseRowTable extends KDatabaseRowAbstract
 	public function __call($method, array $arguments)
 	{ 
 	    // If the method hasn't been mixed yet, load all the behaviors.
-		if(!isset($this->_mixed_methods[$method]))
+	    if($this->isConnected() && !isset($this->_mixed_methods[$method]))
 		{
 			foreach($this->getTable()->getBehaviors() as $behavior) {
 				$this->mixin($behavior);
